Move postprocess to runner
